+++ LMeter/ACT/IpcClient.cs
@@ -21,7 +21,7 @@ namespace LMeter.Act
 
         public IpcClient(ActConfig config) : base(config)
         {
-            _subscriptionReceiver = Singletons.Get<DalamudPluginInterface>().GetIpcProvider<JObject, bool>(LMeterSubscriptionIpcEndpoint);
+            _subscriptionReceiver = Singletons.Get<IDalamudPluginInterface>().GetIpcProvider<JObject, bool>(LMeterSubscriptionIpcEndpoint);
             _subscriptionReceiver.RegisterFunc(ReceiveIpcMessage);
         }
 
@@ -46,7 +46,7 @@ namespace LMeter.Act
             try
             {
                 this.Status = ConnectionStatus.Connecting;
-                bool result = Singletons.Get<DalamudPluginInterface>()
+                bool result = Singletons.Get<IDalamudPluginInterface>()
                     .GetIpcSubscriber<bool>(IinactListeningIpcEndpoint).InvokeFunc();
 
                 Singletons.Get<IPluginLog>().Info("Check if IINACT installed and running: " + result);
@@ -66,7 +66,7 @@ namespace LMeter.Act
             Singletons.Get<IPluginLog>().Info("Successfully discovered IINACT IPC endpoint");
             try
             {
-                bool result = Singletons.Get<DalamudPluginInterface>()
+                bool result = Singletons.Get<IDalamudPluginInterface>()
                     .GetIpcSubscriber<string, bool>(IinactSubscribeIpcEndpoint)
                     .InvokeFunc(LMeterSubscriptionIpcEndpoint);
 
@@ -85,7 +85,7 @@ namespace LMeter.Act
 
             try
             {
-                Singletons.Get<DalamudPluginInterface>()
+                Singletons.Get<IDalamudPluginInterface>()
                     .GetIpcSubscriber<JObject, bool>(IinactProviderEditEndpoint)
                     .InvokeAction(SubscriptionMessageObject);
 
@@ -126,7 +126,7 @@ namespace LMeter.Act
 
             try
             {
-                bool result = Singletons.Get<DalamudPluginInterface>()
+                bool result = Singletons.Get<IDalamudPluginInterface>()
                     .GetIpcSubscriber<string, bool>(IinactUnsubscribeIpcEndpoint)
                     .InvokeFunc(LMeterSubscriptionIpcEndpoint);